+++ plume/src/main/kotlin/io/github/plume/oss/domain/mappers/VertexMapper.kt
@@ -75,9 +75,6 @@ object VertexMapper {
             }
         }
         return when (map["label"] as String) {
-            Binding.Label() -> NewBindingBuilder()
-                .name(map[NAME] as String)
-                .signature(map[SIGNATURE] as String)
             MetaData.Label() -> NewMetaDataBuilder()
                 .language(map[LANGUAGE] as String)
                 .version(map[VERSION] as String)
@@ -250,7 +247,6 @@ object VertexMapper {
      */
     fun stripUnusedProperties(label: String, map: MutableMap<String, Any>): MutableMap<String, Any> {
         return when (label) {
-            BINDING -> map.apply { remove(IS_METHOD_NEVER_OVERRIDDEN) }
             CONTROL_STRUCTURE ->  map.apply { remove(PARSER_TYPE_NAME) }
             JUMP_TARGET ->  map.apply { remove(PARSER_TYPE_NAME) }
             METHOD_REF -> map.apply { remove(TYPE_FULL_NAME) }
@@ -281,7 +277,6 @@ object VertexMapper {
      */
     fun checkSchemaConstraints(fromLabel: String, toLabel: String, edge: String, silent: Boolean = false): Boolean {
         val outRule = when (fromLabel) {
-            Binding.Label() -> Binding.`Edges$`.`MODULE$`.Out().contains(edge)
             MetaData.Label() -> MetaData.`Edges$`.`MODULE$`.Out().contains(edge)
             File.Label() -> File.`Edges$`.`MODULE$`.Out().contains(edge)
             Method.Label() -> Method.`Edges$`.`MODULE$`.Out().contains(edge)
@@ -313,7 +308,6 @@ object VertexMapper {
             }
         }
         val toRule = when (toLabel) {
-            Binding.Label() -> Binding.`Edges$`.`MODULE$`.In().contains(edge)
             MetaData.Label() -> MetaData.`Edges$`.`MODULE$`.In().contains(edge)
             File.Label() -> File.`Edges$`.`MODULE$`.In().contains(edge)
             Method.Label() -> Method.`Edges$`.`MODULE$`.In().contains(edge)