+++ plume/src/test/kotlin/io/github/plume/oss/drivers/TigerGraphDriverIntTest.kt
@@ -5,7 +5,6 @@ import io.github.plume.oss.TestDomainResources.Companion.BOOL_1
 import io.github.plume.oss.TestDomainResources.Companion.INT_1
 import io.github.plume.oss.TestDomainResources.Companion.STRING_1
 import io.github.plume.oss.TestDomainResources.Companion.STRING_2
-import io.github.plume.oss.TestDomainResources.Companion.bindingVertex
 import io.github.plume.oss.TestDomainResources.Companion.blockVertex
 import io.github.plume.oss.TestDomainResources.Companion.callVertex
 import io.github.plume.oss.TestDomainResources.Companion.controlStructureVertex
@@ -84,20 +83,6 @@ class TigerGraphDriverIntTest {
     @Nested
     @DisplayName("Test driver vertex find and exist methods")
     inner class VertexAddAndExistsTests {
-
-        @Test
-        fun findBindingVertex() {
-            val v1 = NewBindingBuilder().name(STRING_1).signature(STRING_2)
-            val v2 = NewBindingBuilder().name(STRING_2).signature(STRING_1)
-            assertFalse(driver.exists(v1))
-            assertFalse(driver.exists(v2))
-            driver.addVertex(v1)
-            assertTrue(driver.exists(v1))
-            assertFalse(driver.exists(v2))
-            driver.addVertex(v2)
-            assertTrue(driver.exists(v1))
-            assertTrue(driver.exists(v2))
-        }
 
         @Test
         fun findFieldIdentifierVertex() {
@@ -292,15 +277,6 @@ class TigerGraphDriverIntTest {
         }
 
         @Test
-        fun testRefEdgeCreation() {
-            assertFalse(driver.exists(bindingVertex, methodVertex, REF))
-            driver.addEdge(bindingVertex, methodVertex, REF)
-            assertTrue(driver.exists(bindingVertex))
-            assertTrue(driver.exists(methodVertex))
-            assertTrue(driver.exists(bindingVertex, methodVertex, REF))
-        }
-
-        @Test
         fun testReceiverEdgeCreation() {
             assertFalse(driver.exists(callVertex, identifierVertex, RECEIVER))
             driver.addEdge(callVertex, identifierVertex, RECEIVER)
@@ -320,11 +296,11 @@ class TigerGraphDriverIntTest {
 
         @Test
         fun testBindsEdgeCreation() {
-            assertFalse(driver.exists(typeDeclVertex, bindingVertex, BINDS))
-            driver.addEdge(typeDeclVertex, bindingVertex, BINDS)
-            assertTrue(driver.exists(typeDeclVertex))
-            assertTrue(driver.exists(bindingVertex))
-            assertTrue(driver.exists(typeDeclVertex, bindingVertex, BINDS))
+            assertFalse(JanusGraphDriverIntTest.driver.exists(typeDeclVertex, methodVertex, AST))
+            JanusGraphDriverIntTest.driver.addEdge(typeDeclVertex, methodVertex, AST)
+            assertTrue(JanusGraphDriverIntTest.driver.exists(typeDeclVertex))
+            assertTrue(JanusGraphDriverIntTest.driver.exists(methodVertex))
+            assertTrue(JanusGraphDriverIntTest.driver.exists(typeDeclVertex, methodVertex, AST))
         }
 
         @Test
@@ -515,7 +491,7 @@ class TigerGraphDriverIntTest {
 
         @Test
         fun testGetProgramStructure() {
-            val unknown = io.shiftleft.semanticcpg.language.types.structure.File.UNKNOWN()
+            val unknown = io.shiftleft.semanticcpg.language.types.structure.FileTraversal.UNKNOWN()
             driver.addVertex(NewFileBuilder().name(unknown).order(0).hash(Option.apply(unknown)))
             g = driver.getProgramStructure()
             val ns = g.nodes().asSequence().toList()
@@ -609,14 +585,14 @@ class TigerGraphDriverIntTest {
         fun testGetIdInsideRange() {
             val ids1 = driver.getVertexIds(0, 10)
             assertTrue(ids1.isEmpty())
-            driver.addVertex(NewBindingBuilder().name(STRING_1).signature(STRING_2).id(1L))
+            driver.addVertex(NewNamespaceBuilder().name(STRING_1).code(STRING_2).order(INT_1).id(1L))
             val ids2 = driver.getVertexIds(0, 10)
             assertEquals(setOf(1L), ids2)
         }
 
         @Test
         fun testGetIdOutsideRange() {
-            driver.addVertex(NewBindingBuilder().name(STRING_1).signature(STRING_2).id(11L))
+            driver.addVertex(NewNamespaceBuilder().name(STRING_1).code(STRING_2).order(INT_1).id(11L))
             val ids1 = driver.getVertexIds(0, 10)
             assertEquals(emptySet<Long>(), ids1)
         }